backend/requirements.txt
flask==2.3.2              # Flask для создания веб-сервера
flask-cors==3.1.0         # Для решения проблем с CORS
requests==2.31.0          # Для выполнения HTTP-запросов к Nextcloud API
python-dotenv==1.0.0      # Для работы с .env файлами и переменными окружения
xmltodict==0.13.0         # Для парсинга XML в словарь (если нужен для работы с XML)
gunicorn==21.0.0          # Для запуска приложения на продакшене
